updated to remove previous pages
