+++ client/src/App.js
@@ -11,8 +11,7 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
 import Signup from './pages/Signup';
 import Login from './pages/Login';
-import SingleThought from './pages/SingleThought';
-import Profile from './pages/Profile';
+import Team from './pages/Team';
 import Header from './components/Header';
 import Footer from './components/Footer';
 
@@ -61,16 +60,8 @@ function App() {
                 element={<Signup />}
               />
               <Route 
-                path="/me" 
-                element={<Profile />}
-              />
-              <Route 
-                path="/profiles/:username" 
-                element={<Profile />}
-              />
-              <Route 
-                path="/thoughts/:thoughtId"
-                element={<SingleThought />}
+                path="/team/:username" 
+                element={<Team />}
               />
             </Routes>
           </div>
