+++ client/src/components/ThoughtForm/index.js
@@ -1,113 +1,71 @@
-import React, { useState } from 'react';
-import { Link } from 'react-router-dom';
-import { useMutation } from '@apollo/client';
+// import React, { useState } from 'react';
+// import { Link } from 'react-router-dom';
+// import { useMutation } from '@apollo/client';
 
-import { ADD_THOUGHT } from '../../utils/mutations';
-import { QUERY_TEAMS, QUERY_ME } from '../../utils/queries';
+// import Auth from '../../utils/auth';
 
-import Auth from '../../utils/auth';
+// const ThoughtForm = () => {
+//   const [thoughtText, setThoughtText] = useState('');
 
-const ThoughtForm = () => {
-  const [thoughtText, setThoughtText] = useState('');
+//   const [characterCount, setCharacterCount] = useState(0);
 
-  const [characterCount, setCharacterCount] = useState(0);
+//   const handleChange = (event) => {
+//     const { name, value } = event.target;
 
-  const [addThought, { error }] = useMutation(ADD_THOUGHT, {
-    update(cache, { data: { addThought } }) {
-      try {
-        const { thoughts } = cache.readQuery({ query: QUERY_TEAMS });
+//     if (name === 'thoughtText' && value.length <= 280) {
+//       setThoughtText(value);
+//       setCharacterCount(value.length);
+//     }
+//   };
 
-        cache.writeQuery({
-          query: QUERY_TEAMS,
-          data: { thoughts: [addThought, ...thoughts] },
-        });
-      } catch (e) {
-        console.error(e);
-      }
+//   return (
+//     <div>
+//       <h3>What's on your techy mind?</h3>
 
-      // update me object's cache
-      const { me } = cache.readQuery({ query: QUERY_ME });
-      cache.writeQuery({
-        query: QUERY_ME,
-        data: { me: { ...me, thoughts: [...me.thoughts, addThought] } },
-      });
-    },
-  });
+//       {Auth.loggedIn() ? (
+//         <>
+//           <p
+//             className={`m-0 ${
+//               characterCount === 280 || error ? 'text-danger' : ''
+//             }`}
+//           >
+//             Character Count: {characterCount}/280
+//           </p>
+//           <form
+//             className="flex-row justify-center justify-space-between-md align-center"
+//             onSubmit={handleFormSubmit}
+//           >
+//             <div className="col-12 col-lg-9">
+//               <textarea
+//                 name="thoughtText"
+//                 placeholder="Here's a new thought..."
+//                 value={thoughtText}
+//                 className="form-input w-100"
+//                 style={{ lineHeight: '1.5', resize: 'vertical' }}
+//                 onChange={handleChange}
+//               ></textarea>
+//             </div>
 
-  const handleFormSubmit = async (event) => {
-    event.preventDefault();
+//             <div className="col-12 col-lg-3">
+//               <button className="btn btn-primary btn-block py-3" type="submit">
+//                 Add Thought
+//               </button>
+//             </div>
+//             {error && (
+//               <div className="col-12 my-3 bg-danger text-white p-3">
+//                 {error.message}
+//               </div>
+//             )}
+//           </form>
+//         </>
+//       ) : (
+//         <p>
+//           You need to be logged in to share your thoughts. Please{' '}
+//           <Link to="/login">login</Link> or <Link to="/signup">signup.</Link>
+//         </p>
+//       )}
+//     </div>
+//   );
+// };
 
-    try {
-      const { data } = await addThought({
-        variables: {
-          thoughtText,
-          thoughtAuthor: Auth.getProfile().data.username,
-        },
-      });
-
-      setThoughtText('');
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  const handleChange = (event) => {
-    const { name, value } = event.target;
-
-    if (name === 'thoughtText' && value.length <= 280) {
-      setThoughtText(value);
-      setCharacterCount(value.length);
-    }
-  };
-
-  return (
-    <div>
-      <h3>What's on your techy mind?</h3>
-
-      {Auth.loggedIn() ? (
-        <>
-          <p
-            className={`m-0 ${
-              characterCount === 280 || error ? 'text-danger' : ''
-            }`}
-          >
-            Character Count: {characterCount}/280
-          </p>
-          <form
-            className="flex-row justify-center justify-space-between-md align-center"
-            onSubmit={handleFormSubmit}
-          >
-            <div className="col-12 col-lg-9">
-              <textarea
-                name="thoughtText"
-                placeholder="Here's a new thought..."
-                value={thoughtText}
-                className="form-input w-100"
-                style={{ lineHeight: '1.5', resize: 'vertical' }}
-                onChange={handleChange}
-              ></textarea>
-            </div>
-
-            <div className="col-12 col-lg-3">
-              <button className="btn btn-primary btn-block py-3" type="submit">
-                Add Thought
-              </button>
-            </div>
-            {error && (
-              <div className="col-12 my-3 bg-danger text-white p-3">
-                {error.message}
-              </div>
-            )}
-          </form>
-        </>
-      ) : (
-        <p>
-          You need to be logged in to share your thoughts. Please{' '}
-          <Link to="/login">login</Link> or <Link to="/signup">signup.</Link>
-        </p>
-      )}
-    </div>
-  );
-};
-
-export default ThoughtForm;
+// export default ThoughtForm;
